refactor: combine block lists
